Revert "Merge pull request #132 from newfold-labs/FIX/PRESS7-375-fix-retry-logic"
This reverts commit 2000f839030cc0faf6b325f36aeff19423a4b315, reversing
changes made to eec827ece85e2d25438bd9f573582205764f0fa2.
