+++ bootstrap.php
@@ -1,7 +1,6 @@
 <?php
 
 use NewfoldLabs\WP\Module\Data\Data;
-use NewfoldLabs\WP\Module\Data\EventQueue\Queues\BatchQueue;
 use NewfoldLabs\WP\Module\Data\Helpers\Encryption;
 use NewfoldLabs\WP\Module\Data\Helpers\Transient;
 use NewfoldLabs\WP\Module\Data\SiteCapabilities;
@@ -28,7 +27,29 @@ if ( ! function_exists( 'nfd_create_event_queue_table' ) ) {
 	 * Create the event queue table
 	 */
 	function nfd_create_event_queue_table() {
-		BatchQueue::create_table();
+		global $wpdb;
+
+		if ( ! function_exists( 'dbDelta' ) ) {
+			require ABSPATH . 'wp-admin/includes/upgrade.php';
+		}
+
+		$wpdb->hide_errors();
+
+		$charset_collate = $wpdb->get_charset_collate();
+
+		$sql = <<<SQL
+CREATE TABLE {$wpdb->prefix}nfd_data_event_queue (
+	id bigint(20) NOT NULL AUTO_INCREMENT,
+	event longtext NOT NULL,
+	attempts tinyint(3) NOT NULL DEFAULT 0,
+	reserved_at datetime DEFAULT NULL,
+	available_at datetime NOT NULL,
+	created_at datetime NOT NULL,
+	PRIMARY KEY (id)
+	) $charset_collate;
+SQL;
+
+		dbDelta( $sql );
 	}
 }
