+++ tests/phpunit/includes/EventManagerTest.php
@@ -157,6 +157,8 @@ class EventManagerTest extends \WP_Mock\Tools\TestCase {
 	 */
 	public function test_send_saved_events_happy_path(): void {
 
+		$this->markTestSkipped( 'Due to an unidentified bug causing events to be resent, we are temporarily disabling retries.' );
+
 		$batch_queue_mock = Mockery::mock( BatchQueue::class );
 
 		\Patchwork\redefine(
@@ -174,13 +176,9 @@ class EventManagerTest extends \WP_Mock\Tools\TestCase {
 
 		$event = Mockery::mock( Event::class );
 
-		$batch_queue_mock->expects( 'remove_events_exceeding_attempts_limit' )
-						->once()
-						->with( 3 );
-
 		$batch_queue_mock->expects( 'pull' )
 						->once()
-						->with( 50 )
+						->with( 100 )
 						->andReturn(
 							array(
 								15 => $event,
@@ -218,10 +216,6 @@ class EventManagerTest extends \WP_Mock\Tools\TestCase {
 		$batch_queue_mock->expects( 'release' )
 						->once()
 						->with( array( 16 ) );
-		
-		$batch_queue_mock->expects( 'increment_attempt' )
-						->once()
-						->with( array( 16 ) );		
 
 		$sut->send_saved_events_batch();
 
@@ -232,6 +226,8 @@ class EventManagerTest extends \WP_Mock\Tools\TestCase {
 	 * @covers ::send_saved_events_batch
 	 */
 	public function test_send_saved_events_happy_path_no_failed_events(): void {
+
+		$this->markTestSkipped( 'Due to an unidentified bug causing events to be resent, we are temporarily disabling retries.' );
 
 		$batch_queue_mock = Mockery::mock( BatchQueue::class );
 
@@ -250,13 +246,9 @@ class EventManagerTest extends \WP_Mock\Tools\TestCase {
 
 		$event = Mockery::mock( Event::class );
 
-		$batch_queue_mock->expects( 'remove_events_exceeding_attempts_limit' )
-						->once()
-						->with( 3 );
-
 		$batch_queue_mock->expects( 'pull' )
 						->once()
-						->with( 50 )
+						->with( 100 )
 						->andReturn(
 							array(
 								15 => $event,
@@ -294,9 +286,6 @@ class EventManagerTest extends \WP_Mock\Tools\TestCase {
 		$batch_queue_mock->expects( 'release' )
 						->never();
 
-		$batch_queue_mock->expects( 'increment_attempt' )
-						->never();
-
 		$sut->send_saved_events_batch();
 
 		$this->assertConditionsMet();
@@ -306,6 +295,8 @@ class EventManagerTest extends \WP_Mock\Tools\TestCase {
 	 * @covers ::send_saved_events_batch
 	 */
 	public function test_send_saved_events_happy_path_no_successful_events(): void {
+
+		$this->markTestSkipped( 'Due to an unidentified bug causing events to be resent, we are temporarily disabling retries.' );
 
 		$batch_queue_mock = Mockery::mock( BatchQueue::class );
 
@@ -323,10 +314,6 @@ class EventManagerTest extends \WP_Mock\Tools\TestCase {
 		$sut = Mockery::mock( EventManager::class )->makePartial();
 
 		$event = Mockery::mock( Event::class );
-
-		$batch_queue_mock->expects( 'remove_events_exceeding_attempts_limit' )
-						->once()
-						->with( 3 );
 
 		$batch_queue_mock->expects( 'pull' )
 						->once()
@@ -364,10 +351,6 @@ class EventManagerTest extends \WP_Mock\Tools\TestCase {
 		$batch_queue_mock->expects( 'remove' )
 						->never();
 
-		$batch_queue_mock->expects( 'increment_attempt' )
-				->once()
-				->with( array( 16 ) );
-
 		$batch_queue_mock->expects( 'release' )
 						->once()
 						->with( array( 16 ) );
@@ -381,6 +364,8 @@ class EventManagerTest extends \WP_Mock\Tools\TestCase {
 	 * @covers ::send_saved_events_batch
 	 */
 	public function test_send_saved_events_wp_error_from_hiive_connection(): void {
+
+		$this->markTestSkipped( 'Due to an unidentified bug causing events to be resent, we are temporarily disabling retries.' );
 
 		$batch_queue_mock = Mockery::mock( BatchQueue::class );
 
@@ -398,10 +383,6 @@ class EventManagerTest extends \WP_Mock\Tools\TestCase {
 		$sut = Mockery::mock( EventManager::class )->makePartial();
 
 		$event = Mockery::mock( Event::class );
-
-		$batch_queue_mock->expects( 'remove_events_exceeding_attempts_limit' )
-		->once()
-		->with( 3 );
 
 		$batch_queue_mock->expects( 'pull' )
 						->once()
@@ -430,8 +411,7 @@ class EventManagerTest extends \WP_Mock\Tools\TestCase {
 		WP_Mock::userFunction( 'is_wp_error' )
 			->once()
 			->andReturnTrue();
-		
-		$batch_queue_mock->expects( 'increment_attempt' )->never();
+
 		$batch_queue_mock->expects( 'remove' )->never();
 
 		$batch_queue_mock->expects( 'release' )
@@ -447,6 +427,8 @@ class EventManagerTest extends \WP_Mock\Tools\TestCase {
 	 * @covers ::send_saved_events_batch
 	 */
 	public function test_send_saved_events_failures_from_hiive(): void {
+
+		$this->markTestSkipped( 'Due to an unidentified bug causing events to be resent, we are temporarily disabling retries.' );
 
 		$batch_queue_mock = Mockery::mock( BatchQueue::class );
 
@@ -464,10 +446,6 @@ class EventManagerTest extends \WP_Mock\Tools\TestCase {
 		$sut = Mockery::mock( EventManager::class )->makePartial();
 
 		$event = Mockery::mock( Event::class );
-		
-		$batch_queue_mock->expects( 'remove_events_exceeding_attempts_limit' )
-						->once()
-						->with( 3 );
 
 		$batch_queue_mock->expects( 'pull' )
 						->once()
@@ -502,10 +480,6 @@ class EventManagerTest extends \WP_Mock\Tools\TestCase {
 			->once()
 			->andReturnFalse();
 
-		$batch_queue_mock->expects( 'increment_attempt' )
-			->once()
-			->with( array( 19 ) );
-
 		$batch_queue_mock->expects( 'release' )
 						->once()
 						->with( array( 19 ) );
@@ -523,12 +497,12 @@ class EventManagerTest extends \WP_Mock\Tools\TestCase {
 	 */
 	public function test_shutdown_happy_path_no_failed_events(): void {
 
+		$this->markTestSkipped( 'Due to an unidentified bug causing events to be resent, we are temporarily disabling retries.' );
+
 		$sut = new EventManager();
 
 		$event      = Mockery::mock( Event::class )->makePartial();
 		$event->key = 'test';
-
-		WP_Mock::expectAction( 'nfd_event_log', $event->key, $event );
 
 		$sut->push( $event );
 
@@ -546,7 +520,7 @@ class EventManagerTest extends \WP_Mock\Tools\TestCase {
 		);
 
 		$hiive_connection_subscriber = Mockery::mock( HiiveConnection::class );
-		WP_Mock::expectFilter('newfold_data_subscribers', $hiive_connection_subscriber);
+
 		$sut->add_subscriber( $hiive_connection_subscriber );
 
 		$hiive_connection_subscriber->expects( 'notify' )
@@ -575,12 +549,12 @@ class EventManagerTest extends \WP_Mock\Tools\TestCase {
 	 */
 	public function test_shutdown_happy_path_with_failed_events(): void {
 
+		$this->markTestSkipped( 'Due to an unidentified bug causing events to be resent, we are temporarily disabling retries.' );
+
 		$sut = new EventManager();
 
 		$event      = Mockery::mock( Event::class )->makePartial();
 		$event->key = 'test';
-		
-		WP_Mock::expectAction( 'nfd_event_log', $event->key, $event );
 
 		$sut->push( $event );
 
@@ -598,8 +572,6 @@ class EventManagerTest extends \WP_Mock\Tools\TestCase {
 		);
 
 		$hiive_connection_subscriber = Mockery::mock( HiiveConnection::class );
-
-		WP_Mock::expectFilter('newfold_data_subscribers', $hiive_connection_subscriber);
 
 		$sut->add_subscriber( $hiive_connection_subscriber );
 
@@ -630,12 +602,12 @@ class EventManagerTest extends \WP_Mock\Tools\TestCase {
 	 */
 	public function test_shutdown_hiive_connection_wp_error(): void {
 
+		$this->markTestSkipped( 'Due to an unidentified bug causing events to be resent, we are temporarily disabling retries.' );
+
 		$sut = new EventManager();
 
 		$event      = Mockery::mock( Event::class )->makePartial();
 		$event->key = 'test';
-
-		WP_Mock::expectAction('nfd_event_log', $event->key, $event);
 
 		$sut->push( $event );
 
@@ -653,7 +625,7 @@ class EventManagerTest extends \WP_Mock\Tools\TestCase {
 		);
 
 		$hiive_connection_subscriber = Mockery::mock( HiiveConnection::class );
-		WP_Mock::expectFilter('newfold_data_subscribers', $hiive_connection_subscriber);
+
 		$sut->add_subscriber( $hiive_connection_subscriber );
 
 		$hiive_connection_subscriber->expects( 'notify' )
@@ -678,12 +650,12 @@ class EventManagerTest extends \WP_Mock\Tools\TestCase {
 	 */
 	public function test_shutdown_hiive_500_error(): void {
 
+		$this->markTestSkipped( 'Due to an unidentified bug causing events to be resent, we are temporarily disabling retries.' );
+
 		$sut = new EventManager();
 
 		$event      = Mockery::mock( Event::class )->makePartial();
 		$event->key = 'test';
-
-		WP_Mock::expectAction('nfd_event_log', $event->key, $event);
 
 		$sut->push( $event );
 
@@ -703,7 +675,7 @@ class EventManagerTest extends \WP_Mock\Tools\TestCase {
 		$hiive_connection_subscriber = Mockery::mock( HiiveConnection::class );
 
 		$sut->add_subscriber( $hiive_connection_subscriber );
-		WP_Mock::expectFilter('newfold_data_subscribers', $hiive_connection_subscriber);
+
 		$hiive_connection_subscriber->expects( 'notify' )
 									->once()
 									->andReturn(
@@ -746,10 +718,6 @@ class EventManagerTest extends \WP_Mock\Tools\TestCase {
 		$sut = Mockery::mock( EventManager::class )->makePartial();
 
 		$event = Mockery::mock( Event::class );
- 
-		$batch_queue_mock->expects( 'remove_events_exceeding_attempts_limit' )
-						->once()
-						->with( 3 );
 
 		$batch_queue_mock->expects( 'pull' )
 						->once()
@@ -771,5 +739,4 @@ class EventManagerTest extends \WP_Mock\Tools\TestCase {
 
 		$this->assertConditionsMet();
 	}
-
 }
